Remove Scala braces to adhere to the style guide
Scala if statements can have braces, but convention is not to use them
except in specific situations; so let's not use them by default.

Let's also add a space before the opening `(`, as this is also
convention.

Lastly, remove the useless placeholder text, as recommende in the
project README. Everyone knows how an if statement works. :)

http://www.codecommit.com/scala-style-guide.pdf