Suggest a better default for Perl package names
The previous snippet for Perl 'package' was apparently copy-pasted from
html.snippets (see 'head' and 'title' snippets). It didn't really make
sense, and for me it actually always returned an empty string.

This one looks for the path name of the file and removes everything up
to the lib/ folder. It then replaces '/' with '::'. I think it's a
pretty sensible default.

Finally, I also added "use strict" and "use warnings" to the package
definition.