Make ---/=== markdown snippets find headline contents
As of version 5352d98f of UltiSnips (not tested with the other snippet
engines), when executing these two snippets the cursor is placed at the
line where `${0}` is located, meaning that when `line('.')` is
evaluated, the line returned is not (as the snippets currently expect)
the line where the `---` or `===` was present, but two lines further
down.

It appears that the order of evaluation is such that first the parts of
the snippet that require no interpolation get written out, then the
inlined code gets executed. As explained above, this leads to
`line('.')` returning the line number of an empty line.

The subtraction statements currently expect to have '.' return the line
where the `---` or `===` is locted. Ensure that this is the case.