Elixir: Fix mirror of `defsd`, add `defs`
Mirror is not using braces, so it's `$1` not `${1}` (at least for UltiSnips, but SnipMate supports this syntax too https://github.com/garbas/vim-snipmate/blob/master/doc/SnipMate.txt#L379).

Seems like this is a feature and not a bug of UltiSnips (https://github.com/SirVer/ultisnips/issues/875, https://github.com/SirVer/ultisnips/issues/875).

I have also added a `defs` which hasn't got the docstring, but is otherwise the same as `defsd`. Perhaps I should always use a docstring, but sometimes I feel the name of the function is descriptive enough.