Speed up the startup time
Now git is called only 2 times and using `command` makes the shell avoid resolving aliases. That decreases the startup time of the plugin by around 300ms on my laptop.
Anyways IMHO those vars should be on the README, because, even with my fix, this plugin makes vim to take noticeable time to start if they are not set.