Do not call external processes during plugin initialization
Having the plugin initialization calling Git commands increases the
startup time by orders of magnitude.

The plugin was very fast before (output of --startuptime):

    000.047  000.047: sourcing .../vim-snippets/plugin/vimsnippets.vim

And this is the result of adding those external Git command calls:

    039.083  039.083: sourcing .../vim-snippets/plugin/vimsnippets.vim

It could be argued that it is still fast, but a lot of users care about
Vim startup time and those 40 ms in this case are too much, and others
have reported worst. In my case, it is taking a fifth of the total
startup time now with almost 50 plugins, and it takes more than 2.5x
longer than any other. Specially when the plugin initialization was
barely noticed originally. Having `system()` calls on a
`<rtp>/plugin/*.vim` file is just bad for the users, when a plugin
should try to load as fast as possible, lazy-loading as much as
possible.

This commit reverts `plugin/vimsnippets.vim` to the state it was before
PR #1252, effectively reverting the PR and all the subsequent commits
trying to fix it:

- 730443a (trim() not exist in old version of Vim., 2020-09-26)
- a044188 (Remove command, 2020-09-25)
- db91ad4 (Update logic using https://github.com/honza/vim-snippets/issues/1257, 2020-09-25)
- 06fbadd (Speed up the startup time, 2020-09-24)
- 5b5cc36 (Correct syntax, 2020-09-23)
- d4c285e (Set author, email and GitHub address automatically from git config., 2020-09-19)

The variables `g:snips_author` and `g:snips_author_email` are not really
widely used to justify having them set automatically from Git at the
cost of increasing the startup time (the workaround for this was having
to op-out of it by defining them in the .vimrc, but that's still not
ideal compared to the opt-in behaviour restored by this commit). Also,
they have been in use with empty/dummy default values since the initial
commit almost 10 years ago, without being an inconvenience to users.

Fixes #1267.