Qualifying Range-v3 and C++20's <ranges> entities; ad-hoc snippet for transform and filter
I find myself very frequently writing `ranges::views::`, and I'd be doing the same
with `std::views::` if I was using C++20, so I think this snippets can be useful.

`transform` and `filter` are two of the most used adapters, so haveing a snippet
for them is good (for both `std::` and `ranges::`, at the price of a backspace).